chore(batch): sync messages script and parse atlas column defaults
Point sync:messages at copy-messages-to-public.ts for docs/specs output.
Expose sql() defaults in the batch DB docs parser.

Made-with: Cursor
